stlink: fix alignment build warning
The {read,write}_mem32 interface functions was asking a 32 bits buffer
but they don't need 32 bits alignment.
This will change the interface to a 8 bits buffer to remove the
alignment mismatch warning. This was causing build errors on platforms
with strict aliasing rules.

Reviewed-on: http://openocd.zylin.com/483
Tested-by: jenkins

+++ src/jtag/stlink/stlink_layout.h
@@ -56,10 +56,10 @@ struct stlink_layout_api_s {
 			    const uint8_t *buffer);
 	/** */
 	int (*read_mem32) (void *handle, uint32_t addr, uint16_t len,
-			   uint32_t *buffer);
+			   uint8_t *buffer);
 	/** */
 	int (*write_mem32) (void *handle, uint32_t addr, uint16_t len,
-			    const uint32_t *buffer);
+			    const uint8_t *buffer);
 	/** */
 	int (*idcode) (void *fd, uint32_t *idcode);
 	/** */
